move to env command
